pyproject.toml
[project]
name = "cyber-wingman"
version = "0.1.0"
description = "赛博僚机 — 两性情感 AI 助手 Agent 后端"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "litellm>=1.81.5,<2.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "loguru>=0.7.3,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "sse-starlette>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-multipart>=0.0.20",
    "duckduckgo-search>=8.1.1",
    "tavily-python>=0.7.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cyber_wingman"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
